Move state back out of it's own lil module
